Restructure and beautify
Breaks Geocoder and each of the the services out of __init__.py into
their own files.  Also includes changes (mostly whitespace) to resolve a
bunch of linter warnings.
